1906. Minimum Absolute Difference Queries
