+++ src/main/kotlin/me/bossm0n5t3r/leetcode/minimumAbsoluteDifferenceQueries/README.md
